+++ src/components/drag-and-drop/custom-drag-layer/custom-drag-layer.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames';
 import { DragLayer } from 'react-dnd';
 
 const collect = (monitor) => {
@@ -62,6 +63,10 @@ class CustomDragLayer extends React.Component {
     };
   }
 
+  getClassName = (props) => {
+    return classNames('custom-drag-layer', props.className);
+  }
+
   setClonedChildWidth(props) {
     this.width = props.draggableNode.getBoundingClientRect().width;
   }
@@ -82,7 +87,7 @@ class CustomDragLayer extends React.Component {
 
   render() {
     return (
-      <div className='custom-drag-layer'>
+      <div className={ this.getClassName(this.props) }>
         <div
           className='custom-drag-layer__container'
           ref={ (node) => { this._container = node; } }