+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Improvements
 
 * `Menu` now outputs semantic HTML. Links are rendered in HTML lists, with submenus rendered with nested lists.
+* `CustomDragLayer` component can now take className prop. 
+
+## Upgrades
+
+* Upgraded carbon-factory to v4.2.2 to support Window environments.
 
 ## Bug Fixes